add failing tests
